+++ src/inspection/inspection.ts
@@ -6,9 +6,9 @@ import { TriedTraverse } from "../common/traversal";
 import { NodeIdMap, ParseError } from "../parser";
 import { InspectionSettings } from "../settings";
 import { ActiveNode, ActiveNodeUtils } from "./activeNode";
-import { AutocompleteInspected, tryFrom as autocompleteInspectedTryFrom } from "./autocomplete";
-import { IdentifierInspected, tryFrom as identifierInspectedTryFrom } from "./identifier";
+import { InspectedAutocomplete, tryFrom as autocompleteInspectedTryFrom } from "./autocomplete";
 import { Position } from "./position";
+import { InspectedScope, tryFrom as scopeInspectedTryFrom } from "./scope";
 
 // Inspection is designed to run sub-inspections,
 // eg. one inspection for scope and one for keywords.
@@ -19,7 +19,7 @@ import { Position } from "./position";
 export interface InspectedCommon {
     readonly maybeActiveNode: ActiveNode | undefined;
 }
-export type Inspected = InspectedCommon & IdentifierInspected & AutocompleteInspected;
+export type Inspected = InspectedCommon & InspectedScope & InspectedAutocomplete;
 export type TriedInspection = Result<Inspected, CommonError.CommonError>;
 
 export function tryFrom(
@@ -35,17 +35,17 @@ export function tryFrom(
         leafNodeIds,
     );
 
-    const triedInspectedIdentifier: TriedTraverse<IdentifierInspected> = identifierInspectedTryFrom(
+    const triedInspectedScope: TriedTraverse<InspectedScope> = scopeInspectedTryFrom(
         settings,
         maybeActiveNode,
         nodeIdMapCollection,
         leafNodeIds,
     );
-    if (ResultUtils.isErr(triedInspectedIdentifier)) {
-        return triedInspectedIdentifier;
+    if (ResultUtils.isErr(triedInspectedScope)) {
+        return triedInspectedScope;
     }
 
-    const triedInspectedKeyword: TriedTraverse<AutocompleteInspected> = autocompleteInspectedTryFrom(
+    const triedInspectedKeyword: TriedTraverse<InspectedAutocomplete> = autocompleteInspectedTryFrom(
         settings,
         maybeActiveNode,
         nodeIdMapCollection,
@@ -57,7 +57,7 @@ export function tryFrom(
 
     return ResultUtils.okFactory({
         maybeActiveNode,
-        ...triedInspectedIdentifier.value,
+        ...triedInspectedScope.value,
         ...triedInspectedKeyword.value,
     });
 }
