+++ src/test/inspection/autocomplete.ts
@@ -8,7 +8,7 @@ import { ResultUtils } from "../../common";
 import { Inspected } from "../../inspection";
 import { KeywordKind, TExpressionKeywords } from "../../lexer";
 import { Ast } from "../../parser";
-import { expectParseErrInspection, expectParseOkInspection, expectTextWithPosition } from "./common";
+import { expectParseErrInspection, expectParseOkInspection, expectTextWithPosition } from "../common";
 
 type AbridgedInspection = Inspected["autocompleteKeywords"];
